requirements.txt
requests == 2.27.1
pyyaml
schedule
APScheduler